move to shared-styles, and reorder imports
