+++ context/app/static/js/components/Detail/Attribution/Attribution.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import { LightBlueLink } from 'js/shared-styles/Links';
+import SectionHeader from 'js/shared-styles/sections/SectionHeader';
+import SectionContainer from 'js/shared-styles/sections/SectionContainer';
 import { FlexPaper } from './style';
-import SectionHeader from '../SectionHeader';
-import SectionContainer from '../SectionContainer';
 import SectionItem from '../SectionItem';
 
 function Attribution(props) {
