+++ context/app/static/js/components/Detail/SampleTissue/SampleTissue.jsx
@@ -2,10 +2,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import { LightBlueLink } from 'js/shared-styles/Links';
+import SectionHeader from 'js/shared-styles/sections/SectionHeader';
+import SectionContainer from 'js/shared-styles/sections/SectionContainer';
 import { FlexPaper } from './style';
-import SectionHeader from '../SectionHeader';
 import SectionItem from '../SectionItem';
-import SectionContainer from '../SectionContainer';
 
 function MetadataItem(props) {
   const { label, children, ml, flexBasis } = props;
