+++ context/app/static/js/components/files/Files/Files.jsx
@@ -2,8 +2,8 @@ import React, { useState, useContext } from 'react';
 import PropTypes from 'prop-types';
 
 import DetailContext from 'js/components/Detail/context';
-import SectionHeader from 'js/components/Detail/SectionHeader';
-import SectionContainer from 'js/components/Detail/SectionContainer';
+import SectionHeader from 'js/shared-styles/sections/SectionHeader';
+import SectionContainer from 'js/shared-styles/sections/SectionContainer';
 import GlobusLink from '../GlobusLink';
 import FileBrowser from '../FileBrowser';
 import FileBrowserDUA from '../FileBrowserDUA';
