+++ context/app/static/js/pages/Collections/Collections.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useContext } from 'react';
 import Typography from '@material-ui/core/Typography';
 
 import { AppContext } from 'js/components/Providers';
-import Description from 'js/components/Detail/Description';
+import Description from 'js/shared-styles/sections/Description';
 import Panel from 'js/components/Collections/Panel';
 import { PageWrapper, ScrollBox } from './style';
