+++ context/app/static/js/pages/Preview/Preview.jsx
@@ -3,10 +3,10 @@ import Typography from '@material-ui/core/Typography';
 
 import Markdown from 'js/components/Markdown';
 import VisualizationWrapper from 'js/components/Detail/visualization/VisualizationWrapper';
-import SectionHeader from 'js/components/Detail/SectionHeader';
-import SectionContainer from 'js/components/Detail/SectionContainer';
+import SectionHeader from 'js/shared-styles/sections/SectionHeader';
+import SectionContainer from 'js/shared-styles/sections/SectionContainer';
 import Attribution from 'js/components/Detail/Attribution';
-import Description from 'js/components/Detail/Description';
+import Description from 'js/shared-styles/sections/Description';
 
 function Preview(props) {
   const { vitData, title, assayMetadata, markdown } = props;
