+++ context/app/static/js/pages/Services/Services.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import Paper from '@material-ui/core/Paper';
 
-import SectionHeader from 'js/components/Detail/SectionHeader';
-import SectionContainer from 'js/components/Detail/SectionContainer';
-import Description from 'js/components/Detail/Description';
+import SectionHeader from 'js/shared-styles/sections/SectionHeader';
+import SectionContainer from 'js/shared-styles/sections/SectionContainer';
+import Description from 'js/shared-styles/sections/Description';
 import ServiceStatusTable from 'js/components/ServiceStatusTable';
 
 function Services() {
