+++ context/app/static/js/shared-styles/sections/SectionContainer.jsx
