+++ context/app/static/js/shared-styles/sections/SectionHeader.jsx
